Explicitly define option values
This change specifies an explicit integer value for each option.
This should improve compatibility with clients that use the
dynamically linked library (e.g. TidyManaged). Otherwise when new
options get inserted the following options get new values which the
client doesn't know, so it tries to access it with the old value
which might fail because that value now is a different option with
probably a different type.