Add setting the CMAKE_BUILD_TYPE to build-me.sh.
The default MUST set -DCMAKE_BUILD_TYPE=Release - without this addition
the -DNDEBUG compiler flag is NOT set, leaving 'asserts' in the code.

Asserts are good and necessary for a Debug version, for testing, but
with their brutal abort on error should NOT really be in Release code
IMHO.

And after all this is ONLY in a 'sample' build-me.sh script. Users
are free to use whatever options they desire in their own build of
tidy, and frequently do.