Build both static and shared libraries.
First many other projects do this, and it seems a good choice.

Due mainly to windows where the static and shared libraries have the same
name, appended an 's' to the static library.

The tidy console app will use the static library, since to build the man
page tidy may need to be run before the shared libraries are installed.
And this makes the windows tidy exe a stand-alone app. No DLL install
issue.

This addresses issue #194, and possibly #190