Small fix in platform.h, just for windows (WIN32)
Not sure how this has lasted so long, but I guess there are not too many
users wanting to specifically use the static tidy library.

Now it is possible to define TIDY_STATIC and the TIDY_EXPORT macro will
just be 'external', suitable for linking with the static library, instead
of adding the _imp_ MSVC __declspec( dllimport )...

This fix does not effect other than the WIN32 platform build.