Make internal and extenal form for opaque types equivalent
Now both internal and external types are the same.  The only distinction is
the fact that the definitions of the structures in question are not
available in the public API, so they remain opaque.  The conversion macros
have become no-ops, so the compiler will help checking types.

TidyIterator doesn't actually name a single structure.  Often it's not a
pointer at all, but an integer instead.  So that will still be dealt with as
before.

The TidyOption definition is somewhat special as well, since it contains the
const qualifier, and the name of the structure doesn't follow the common
schema.  For that reason it doesn't make use of the opaque_type macro.

This change was created using a lot of automated text replacements, combined
with some manual editing.  In the next step the conversion macros will be
removed altogether.