Clarify the terms used for both kinds of options
I removed references to "standard" (or regular) command-line options
and "extended" (or detailed) options. I used the terms featured in
the description of the options which output XML files describing
each kind of options:

  -xml-help
        list the command line options in XML format

  -xml-config
        list all configuration options in XML format

The term for single-dash options is now (purely) command-line options
while double-dash options are referred to as configuration options.

Related issue: #469