Clarify configuration options equivalent to command-line options
I added a paragraph to explain the equivalence of a command-line
option with a configuration option and value, and to explicit the
format used to describe this equivalence in the description of
command-line parameters.

I moved the parentheses, which were on the last line, at the end
of the description, to the first line at the end of the list of
names for the command-line option.

Related issue: #469