Wait until after reading all the init files and paramerers before calling AdjustConfig
Calling AdjustConfig right after reading an init file tends to break things since
tidy might not have the all of the config settings yet.  It needs to wait until
all the init files and parameters have been processed so that it can make it's
adjustments based on the full config.