Set the indent character in a more appropriate spot.
The previous location was fine when saving the output stream, but it led to
unset character when using node interrogation, which _also_ uses the pretty
printers to get text. This is a more appropriate location to set the character.

This is the first instance of our cool Swift wrapper API tester catching an
issue for us. Nice!

Proof that the automated API